# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Plant Ride API",
    "description": "In-plant cab booking, pooling, fleet compliance and cost-centre billing for LHS.\n\n1. Call **POST /api/auth/login** with a demo account (password `Plant@123`):\n   `LHS-40218` (rider), `DRV-2281` (driver), `admin` (desk + admin), `desk` (desk).\n2. Click **Authorize** and paste the returned `token`.\n",
    "contact": {
      "name": "SRM ECO TECH",
      "email": "info@srmecotech.com"
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/driver/trips/{tripId}/no-show": {
      "post": {
        "tags": [
          "4. Driver"
        ],
        "summary": "Mark a rider as a no-show after the configured wait",
        "operationId": "noShow",
        "parameters": [
          {
            "name": "tripId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NoShowRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/DriverQueueResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/driver/trips/{tripId}/board": {
      "post": {
        "tags": [
          "4. Driver"
        ],
        "summary": "Board a rider with their 4-digit OTP",
        "operationId": "board",
        "parameters": [
          {
            "name": "tripId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BoardRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/DriverQueueResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/driver/trips/{tripId}/arrive": {
      "post": {
        "tags": [
          "4. Driver"
        ],
        "summary": "Arrived at the next stop: drops complete, waiting riders are told the cab is here",
        "operationId": "arrive",
        "parameters": [
          {
            "name": "tripId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArriveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/DriverQueueResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/driver/sos": {
      "post": {
        "tags": [
          "4. Driver"
        ],
        "summary": "SOS: alerts the transport desk and plant security with the cab's location",
        "operationId": "sos_1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SosRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/IssueResultDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/driver/location": {
      "post": {
        "tags": [
          "4. Driver"
        ],
        "summary": "GPS ping from the driver app: live position, overspeed check, geofence arrival",
        "operationId": "location",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LocationPingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/LocationAckDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/driver/issues": {
      "post": {
        "tags": [
          "4. Driver"
        ],
        "summary": "Report a breakdown, accident, delay, fuel entry or other issue",
        "operationId": "reportIssue",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportIssueRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/IssueResultDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/driver/duty/start": {
      "post": {
        "tags": [
          "4. Driver"
        ],
        "summary": "Sign on: vehicle check, odometer, compliance gate",
        "operationId": "startDuty",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartDutyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/DutyDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/driver/duty/end": {
      "post": {
        "tags": [
          "4. Driver"
        ],
        "summary": "Sign off: closing odometer, fuel, GPS-vs-odometer check",
        "operationId": "endDuty",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EndDutyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/DutySummaryDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/billing/summary": {
      "get": {
        "tags": [
          "6. Billing (admin)"
        ],
        "summary": "Charges by cost centre, vendor bill vs GPS, and compliance watch for a month",
        "operationId": "summary",
        "parameters": [
          {
            "name": "period",
            "in": "query",
            "description": "Billing month yyyy-MM; defaults to the current month",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BillingSummaryDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "RouteRef": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "code": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "label": {
            "type": "string"
          }
        }
      },
...
      "VehicleRef": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "code": {
            "type": "string"
          },
          "registrationNo": {
            "type": "string"
          },
          "vehicleType": {
            "type": "string",
            "enum": [
              "CAB",
              "SHUTTLE"
            ]
          },
          "seatCapacity": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "NoShowRequest": {
        "type": "object",
        "properties": {
          "bookingId": {
            "type": "integer",
            "format": "int64"
          }
        },
        "required": [
          "bookingId"
        ]
      },
      "DriverQueueResponse": {
        "type": "object",
        "properties": {
          "trip": {
            "$ref": "#/components/schemas/TripQueueDto"
          },
          "notice": {
            "type": "string"
          }
        }
      },
      "QueueRiderDto": {
        "type": "object",
        "properties": {
          "bookingId": {
            "type": "integer",
            "format": "int64"
          },
          "bookingCode": {
            "type": "string"
          },
          "riderName": {
            "type": "string"
          },
          "personnelNo": {
            "type": "string"
          },
          "riderPhone": {
            "type": "string"
          },
          "seats": {
            "type": "integer",
            "format": "int32"
          },
          "status": {
            "type": "string"
          },
          "fromStopName": {
            "type": "string"
          },
          "toStopName": {
            "type": "string"
          },
          "noShowAllowedInSeconds": {
            "type": "integer",
            "format": "int64"
          },
          "otpLocked": {
            "type": "boolean"
          }
        }
      },
      "QueueStopDto": {
        "type": "object",
        "properties": {
          "seq": {
            "type": "integer",
            "format": "int32"
          },
          "stopId": {
            "type": "integer",
            "format": "int64"
          },
          "stopName": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "etaMinutes": {
            "type": "integer",
            "format": "int32"
          },
          "summary": {
            "type": "string"
          },
          "pickups": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QueueRiderDto"
            }
          },
          "drops": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QueueRiderDto"
            }
          }
        }
      },
      "TripQueueDto": {
        "type": "object",
        "properties": {
          "tripId": {
            "type": "integer",
            "format": "int64"
          },
          "tripCode": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "PLANNED",
              "IN_PROGRESS",
              "COMPLETED",
              "CANCELLED"
            ]
          },
          "route": {
            "$ref": "#/components/schemas/RouteRef"
          },
          "exclusive": {
            "type": "boolean"
          },
          "capacity": {
            "type": "integer",
            "format": "int32"
          },
          "onBoard": {
            "type": "integer",
            "format": "int32"
          },
          "seatsLeft": {
            "type": "integer",
            "format": "int32"
          },
          "currentSeq": {
            "type": "integer",
            "format": "int32"
          },
          "currentStopName": {
            "type": "string"
          },
          "nextStop": {
            "$ref": "#/components/schemas/QueueStopDto"
          },
          "canArriveNext": {
            "type": "boolean"
          },
          "arriveBlockedReason": {
            "type": "string"
          },
          "speedLimitKmh": {
            "type": "integer",
            "format": "int32"
          },
          "stops": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QueueStopDto"
            }
          }
        }
      },
      "BoardRequest": {
        "type": "object",
        "properties": {
          "bookingId": {
            "type": "integer",
            "format": "int64"
          },
          "otp": {
            "type": "string",
            "minLength": 1,
            "pattern": "\\d{4}"
          }
        },
        "required": [
          "bookingId",
          "otp"
        ]
      },
      "ArriveRequest": {
        "type": "object",
        "properties": {
          "stopSeq": {
            "type": "integer",
            "format": "int32",
            "minimum": 1
          }
        },
        "required": [
          "stopSeq"
        ]
      },
      "SosRequest": {
        "type": "object",
        "properties": {
          "latitude": {
            "type": "number",
            "format": "double"
          },
          "longitude": {
            "type": "number",
            "format": "double"
          },
          "note": {
            "type": "string",
            "maxLength": 300,
            "minLength": 0
          }
        }
      },
      "IssueResultDto": {
        "type": "object",
        "properties": {
          "alertId": {
            "type": "integer",
            "format": "int64"
          },
          "message": {
            "type": "string"
          },
          "ridersRematched": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "LocationPingRequest": {
        "type": "object",
        "properties": {
          "latitude": {
            "type": "number",
            "format": "double",
            "maximum": 90.0,
            "minimum": -90.0
          },
          "longitude": {
            "type": "number",
            "format": "double",
            "maximum": 180.0,
            "minimum": -180.0
          },
          "speedKmh": {
            "type": "number",
            "format": "double",
            "maximum": 250.0
          },
          "recordedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "latitude",
          "longitude"
        ]
      },
      "LocationAckDto": {
        "type": "object",
        "properties": {
          "overspeed": {
            "type": "boolean"
          },
          "speedLimitKmh": {
            "type": "integer",
            "format": "int32"
          },
          "autoArrivedAt": {
            "type": "string"
          }
        }
      },
      "ReportIssueRequest": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "BREAKDOWN",
              "ACCIDENT",
              "DELAY",
              "FUEL",
              "OTHER"
            ]
          },
          "description": {
            "type": "string",
            "maxLength": 400,
            "minLength": 0
          },
          "litres": {
            "type": "number",
            "maximum": 300.0,
            "minimum": 0.0
          }
        },
        "required": [
          "description",
          "type"
        ]
      },
      "StartDutyRequest": {
        "type": "object",
        "properties": {
          "vehicleId": {
            "type": "integer",
            "format": "int64"
          },
          "startOdometer": {
            "type": "integer",
            "format": "int32",
            "maximum": 9999999
          },
          "checklist": {
            "type": "array",
            "items": {
              "type": "string",
              "minLength": 1
            },
            "minItems": 1
          }
        },
        "required": [
          "checklist",
          "startOdometer",
          "vehicleId"
        ]
      },
      "DutyDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "CLOSED"
            ]
          },
          "vehicle": {
            "$ref": "#/components/schemas/VehicleRef"
          },
          "startedAt": {
            "type": "string",
            "format": "date-time"
          },
          "minutesOnDuty": {
            "type": "integer",
            "format": "int64"
          },
          "maxDutyMinutes": {
            "type": "integer",
            "format": "int32"
          },
          "startOdometer": {
            "type": "integer",
            "format": "int32"
          },
          "tripsCompleted": {
            "type": "integer",
            "format": "int32"
          },
          "ridersCarried": {
            "type": "integer",
            "format": "int32"
          },
          "distanceKm": {
            "type": "number"
          }
        }
      },
      "EndDutyRequest": {
        "type": "object",
        "properties": {
          "endOdometer": {
            "type": "integer",
            "format": "int32",
            "maximum": 9999999
          },
          "fuelLitres": {
            "type": "number",
            "maximum": 300.0,
            "minimum": 0.0
          }
        },
        "required": [
          "endOdometer"
        ]
      },
      "DutySummaryDto": {
        "type": "object",
        "properties": {
          "dutyId": {
            "type": "integer",
            "format": "int64"
          },
          "vehicleCode": {
            "type": "string"
          },
          "startedAt": {
            "type": "string",
            "format": "date-time"
          },
          "endedAt": {
            "type": "string",
            "format": "date-time"
          },
          "dutyMinutes": {
            "type": "integer",
            "format": "int64"
          },
          "tripsCompleted": {
            "type": "integer",
            "format": "int32"
          },
          "ridersCarried": {
            "type": "integer",
            "format": "int32"
          },
          "gpsKm": {
            "type": "number"
          },
          "odometerKm": {
            "type": "integer",
            "format": "int32"
          },
          "odometerCheck": {
            "type": "string"
          },
          "fuelLitres": {
            "type": "number"
          }
        }
      },
...
      "BillingSummaryDto": {
        "type": "object",
        "properties": {
          "period": {
            "type": "string"
          },
          "draft": {
            "type": "boolean"
          },
          "sharingRule": {
            "type": "string"
          },
          "totals": {
            "$ref": "#/components/schemas/Totals"
          },
          "costCentres": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CostCentreCharge"
            }
          },
          "vendors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VendorReconciliation"
            }
          },
          "compliance": {
            "$ref": "#/components/schemas/ComplianceWatch"
          }
        }
      },
      "ComplianceWatch": {
        "type": "object",
        "properties": {
          "documentsExpiringSoon": {
            "type": "integer",
            "format": "int32"
          },
          "expiredDocuments": {
            "type": "integer",
            "format": "int32"
          },
          "blockedVehicles": {
            "type": "integer",
            "format": "int32"
          },
          "blockedDrivers": {
            "type": "integer",
            "format": "int32"
          },
          "fullyCompliantVehicles": {
            "type": "integer",
            "format": "int32"
          },
          "totalVehicles": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "CostCentreCharge": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "trips": {
            "type": "integer",
            "format": "int64"
          },
          "rides": {
            "type": "integer",
            "format": "int64"
          },
          "km": {
            "type": "number"
          },
          "amount": {
            "type": "number"
          },
          "budget": {
            "type": "number"
          },
          "budgetUsedPct": {
            "type": "integer",
            "format": "int32"
          },
          "budgetStatus": {
            "type": "string"
          }
        }
      },
      "Totals": {
        "type": "object",
        "properties": {
          "trips": {
            "type": "integer",
            "format": "int64"
          },
          "rides": {
            "type": "integer",
            "format": "int64"
          },
          "gpsKm": {
            "type": "number"
          },
          "amount": {
            "type": "number"
          },
          "ridersPerTrip": {
            "type": "number"
          }
        }
      },
      "VendorReconciliation": {
        "type": "object",
        "properties": {
          "vendorCode": {
            "type": "string"
          },
          "vendorName": {
            "type": "string"
          },
          "claimedKm": {
            "type": "number"
          },
          "gpsKm": {
            "type": "number"
          },
          "gapKm": {
            "type": "number"
          },
          "gapPct": {
            "type": "number"
          },
          "claimedAmount": {
            "type": "number"
          },
          "gpsAmount": {
            "type": "number"
          },
          "status": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}